Read wordlist from dict instead from files
Instead of reading the wordlists from files they are read
from the dictionary specified in wordlists.py.
Instead of paths subfolder names and file names, keys
from the dictionary are used to access the words.

To clarify this I renamed the function load_file to
load_wordlists.

Please note at this stage there is still an issue with two
test cases in the unit tests. This is an intermediary commit
I make so I am able to switch back and forth between
branches more easily.