Fix test_close_matches
Problem was function getallcategories had a different behaviour than
expected. Previously it was operating on filepaths and the remainder of
the code expected the function to return 'pos/category' in some cases
but return only 'category' when 'pos' was specified.
I fixed it by adding an if statement. I believe a reader can now
understand more easily how the function is expected to operate.