Removes as_valid_path as its not necessary anymore
Since we do not deal with paths anymore this function has became useless
It might make sense to bring it back for the generate_wordlist.py app
if we observe some problems on other systems with that tool.
In any case the actual library does not need to use file paths anymore
making the function useless.

All tests pass now