Add entropy and collision probability estimation.
This provides functions to *estimate* entropy and collision probability of IDs
generated via the generate call. These estimates will be accurate if no
func_words (such as 'uuid/') are used. In the presence of functions, sampling
is used to estimate entropy of the function. As the number of samples increases
this will approach an upper bound on the entropy (which will underestimate
collision probability). If the function is not a typical ID function, then
the upper bound may be a lot higher than the true entropy, and hence collision
probability estimates will be too low.